requirements.txt
torch
torchvision
numpy
pandas
matplotlib
Pillow
scikit-learn
streamlit
plotly
